+++ FLATApp/FLATApp/Api/Api.swift
@@ -7,14 +7,6 @@
 
 import SwiftUI
 
-struct UserData: Decodable, Identifiable {
-    var id: Int
-    var name: String
-    var icon_path: String
-    var applied: Bool
-    var requested: Bool
-}
-
 struct Dummy: Codable{}
 
 enum NetworkError : Error{//失敗した時用
@@ -29,47 +21,7 @@ enum HttpMethod{
 }
 
 
-func searchName(target_name: String,
-              success: @escaping ([UserData]) -> (),
-              failure: @escaping (Error) -> ()
-){ //相手の名前を検索した時にその名前をURLに入れる
-    guard let target_name_encode = target_name.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed) else {
-        return
-    }
-    
-    let req_url = "/v1/user/search?my_id=0&target_name=\(target_name_encode)"
-   
-    Api.util(endpoint: req_url, method: HttpMethod.GET, args: Dummy(), success: {(dictionary: [UserData]) in
-        success(dictionary)
-    }) {(error) in
-        failure(error)
-    }
-}
-
-
-struct IdAndBeacon: Codable{
-    var user_id: Int
-    var uuid: String
-    var major: Int
-    var minor: Int
-    var rssi: Float
-    var distance: Float
-}
-
-func sendBeacon(beacon: IdAndBeacon,
-                success: @escaping ([String:String]) -> (),
-                failure: @escaping (Error) -> ()
-){
-    let req_url = "/v1/user/beacon"
-    Api.util(endpoint: req_url, method: HttpMethod.POST, args: beacon, success: {(msg:[String:String]) in
-        print(msg)
-        success(msg)
-    }) {(error) in
-        failure(error)
-    }
-}
-
-final class Api{
+public final class Api{
     private init(){}
     static var baseUrl = "http://34.68.157.198:8080"
     static let shared = URLSession.shared
